Doing something different this time -- but only to core I marked all the empty statements with //nop I removed unneccessary elses (ie, the if returns or throws) I took out some casts (integral promotions/some didn't need to be there) -- Love, shendaras (2 in one day, w00t)

+++ core/java/src/net/i2p/crypto/DSAEngine.java
@@ -157,7 +157,7 @@ public class DSAEngine {
             k += 512;
         }
         int padbytes = k / 8;
-        int wordlength = (int) (source.length / 4 + padbytes / 4 + 3);
+        int wordlength = source.length / 4 + padbytes / 4 + 3;
         int[] M0 = new int[wordlength];
         int wordcount = 0;
         int x = 0;
